Implement another option to work with history
Previously, the whole history of all authors in the dataset was loaded into memory, and previous messages for each commit were obtained on the fly. Since the dataset grown bigger, it led to huge RAM consumption. So I added an option to preprocess history for each example beforehand: higher disk usage, but lower RAM usage.